requirements.txt
Django
djangorestframework
environ
